validation and santization of comments

+++ deleteComment.php
@@ -1,19 +1,22 @@
 <?php
-
 include 'dbConnect.php';
 
-
-$id = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_STRING);
-
-
-
-// Build the parameterized SQL query and bind sanitized values to the parameters
-$query = "DELETE FROM comments WHERE id=:id";
-$statement = $db->prepare($query);
-
-$statement->bindValue(':id', $id);
-
-// Execute the INSERT prepared statement.
-$statement->execute();
-header("Location: http://localhost:81/movie/commentAdmin.php");
-exit;
+if (isset($_POST['id'])) {
+    $id = filter_input(INPUT_POST, 'id', FILTER_SANITIZE_NUMBER_INT);
+    if ($id !== false) {
+        $query = "DELETE FROM comments WHERE id=:id";
+        $statement = $db->prepare($query);
+        $statement->bindValue(':id', $id, PDO::PARAM_INT);
+        if ($statement->execute()) {
+            header("Location: http://localhost:81/movie/commentAdmin.php");
+            exit;
+        } else {
+            echo "Error deleting comment.";
+        }
+    } else {
+        echo "Invalid comment id.";
+    }
+} else {
+    echo "Comment id is missing.";
+}
+?>
